+++ src/components/dashboard/menu/ListMenu.tsx
@@ -1,6 +1,8 @@
 import { useState, useEffect } from 'react';
 import Collapse from './ListMenuCollapse';
-import Static from './ListMenuStatic';
+import Static from './ListMenuFull';
+import { faClipboardList } from '@fortawesome/free-solid-svg-icons';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 
 export default function ListMenu() {
   const [isView, setView] = useState(false);
@@ -25,8 +27,8 @@ export default function ListMenu() {
   return (
     <div className='mx-5 mb-1 mt-5 h-fit w-auto rounded-xl border border-slate-400 bg-slate-700 pb-1 xl:my-5 xl:ml-5 xl:mr-2 xl:h-auto xl:shrink-0 xl:basis-60'>
       <div className=''>
-        <h2 className='m-2 mx-2 rounded border-4 border-slate-100 bg-blue-100 text-xl'>
-          List: Menu
+        <h2 className='m-2 pr-4 rounded border-4 border-slate-100 bg-blue-100 text-xl text-deep-blue'>
+          <span className='mx-2'><FontAwesomeIcon icon={faClipboardList} /></span>Menu
         </h2>
         {isView ? <Static /> : <Collapse />}
       </div>
